Sketch of how to support other frameworks
This uses an anonymous interface type to define the set of methods needed on the testRunner by main, and introduces a flag as a way to specify the framework type.

Aside from the `GoTests` being an empty example, note that there's no environment var that corresponds to the new flag, and there are no flags corresponding to the config currently specified by env var... unifying these might be nice?