If err != nil, don't discover that path
To enable debugging things like permissions issues that prevent Glob from walking a directory that matches the pattern, Glob calls the callback with paths if there was an error. This currently happens even with file paths that don't fully match the pattern (a zzglob bug). 

Regardless, errors being passed into the callback should cause the path to not be discovered.