package bes: refactor, closer to bb-portal's implementation
PublishBuildToolEventStream is copied verbatim from bb-portal,
and BuildEventHandler and BuildEventChannel interfaces/structs are
created to be compatible with that code.

This also resolves a bug in the earlier WIP implementation where most
event sequence IDs were not acknowledged, resulting in a warning shown
on the Bazel side sending the events.