Filter jest command with file paths on retries
The initial test run command for jest specifies which tests to run by specifying file paths, ensuring jest only needs to load and parse those specific files. However, when retrying failed tests, the filtering is done via a regex (passed into `--testNamePattern`). This forces jest to load and parse all test files to then filter down to the matching tests. On large repos this can be a very slow process.

This commit updates the retry command to accept an optional `{{testExamples}}` placeholder to also filter the retry command to only the files that contain the tests to be retried.

This also fixes a bug in the way that testCase paths are constructed to only include the location in the path if the location is available (and not 0:0).