Fix location of Python .venv dir in bin/setup
The `bin/setup` script installs the python `.venv/` directory in the
wrong parent directory `./internal/runner/testdata/cucumber`, rather
than the root directory of the project.

```
$ ./bin/setup
$ find . -name .venv
./internal/runner/testdata/cucumber/.venv
```

This subsequently causes one of the cucumber tests to fail because it
finds unexpected files in the `internal/runner/testdata/cucumber`
directory.

This change uses `pushd` and `popd` so it's not required to keep track
of the pwd in the script.